refactor(core): modularize cli and harden workspace state
