+++ src/script.test.ts
@@ -121,7 +121,7 @@ describe("generateAppleScript", () => {
     const plan = planLayout({ shell: "npm run dev" });
     const script = generateAppleScript(plan, "/tmp");
 
-    expect(script).toContain("set initial input of cfg to \"npm 'run' 'dev'\\n\"");
+    expect(script).toContain('set initial input of cfg to "npm run dev\\n"');
   });
 
   it("does not use delay for pane initialization when auto-resize is off", () => {
@@ -287,37 +287,34 @@ describe("generateAppleScript", () => {
     expect(script).toContain('set initial input of cfg to ""');
   });
 
-  it("escapes single quotes in initial input commands", () => {
+  it("preserves single quotes in initial input commands", () => {
     const plan = planLayout({ sidebarCommand: "cmd 'arg'" });
     const script = generateAppleScript(plan, "/tmp");
 
-    // Single quotes in args still get POSIX-escaped, then AppleScript-escaped
-    expect(script).toContain("set initial input of cfg to \"cmd ''\\\\''arg'\\\\'''\\n\"");
+    expect(script).toContain('set initial input of cfg to "cmd \'arg\'\\n"');
   });
 
-  it("escapes shell metacharacters in root pane editor command", () => {
-    // $HOME should not expand when typed into the shell
+  it("preserves shell syntax in root pane editor command", () => {
+    // User-authored commands are preserved exactly as typed.
     const plan1 = planLayout({ editor: "vim $HOME" });
     const script1 = generateAppleScript(plan1, "/tmp");
-    // The command name stays unquoted, but the argument is shell-quoted
-    expect(script1).toContain("input text \"vim '$HOME'\" to paneRoot");
+    expect(script1).toContain('input text "vim $HOME" to paneRoot');
 
-    // Backtick command substitution should not expand
+    // Backtick command substitution is preserved too.
     const plan2 = planLayout({ editor: "vim `whoami`" });
     const script2 = generateAppleScript(plan2, "/tmp");
-    expect(script2).toContain("input text \"vim '`whoami`'\" to paneRoot");
+    expect(script2).toContain('input text "vim `whoami`" to paneRoot');
 
-    // $() command substitution should not expand (each word is individually quoted)
+    // Quoted arguments and shell syntax stay intact.
     const plan3 = planLayout({ editor: "vim $(rm -rf /)" });
     const script3 = generateAppleScript(plan3, "/tmp");
-    expect(script3).toContain("input text \"vim '$(rm' '-rf' '/)'\"" + " to paneRoot");
+    expect(script3).toContain('input text "vim $(rm -rf /)" to paneRoot');
   });
 
-  it("escapes single quotes in root pane editor command arguments", () => {
+  it("preserves single quotes in root pane editor command arguments", () => {
     const plan = planLayout({ editor: "cmd 'arg'" });
     const script = generateAppleScript(plan, "/tmp");
-    // Single quotes in the argument are POSIX-escaped, then escapeAppleScript doubles backslashes
-    expect(script).toContain("input text \"cmd ''\\\\''arg'\\\\'''\" to paneRoot");
+    expect(script).toContain('input text "cmd \'arg\'" to paneRoot');
   });
 
   it("leaves plain editor command without arguments unchanged", () => {
@@ -326,18 +323,12 @@ describe("generateAppleScript", () => {
     expect(script).toContain('input text "claude" to paneRoot');
   });
 
-  it("splits space-containing arguments into individually quoted tokens (by design)", () => {
-    // quoteCommand splits on spaces, so a command like `grep "hello world"`
-    // becomes `grep 'hello' 'world'` — NOT `grep 'hello world'`.
-    // This is a known limitation of the key=value config format, which has
-    // no syntax for preserving quoted multi-word arguments.
-    // Each fragment is still POSIX single-quoted, so this is NOT an injection vector.
+  it("preserves quoted multi-word arguments", () => {
     const plan = planLayout({ editor: 'grep "hello world" file.txt' });
     const script = generateAppleScript(plan, "/tmp");
 
-    // Each space-delimited token after the command name is individually quoted
     expect(script).toContain(
-      "input text \"grep '\\\"hello' 'world\\\"' 'file.txt'\" to paneRoot",
+      'input text "grep \\"hello world\\" file.txt" to paneRoot',
     );
   });
 
@@ -380,8 +371,8 @@ describe("generateAppleScript", () => {
     // Right column exists only for shell, and shell has a specific command
     expect(script).toContain("paneRightCol");
     // The shell command should be set via initial input before the right column split
-    expect(script).toContain("set initial input of cfg to \"npm 'run' 'dev'\\n\"");
-    const cmdIndex = script.indexOf("npm 'run' 'dev'");
+    expect(script).toContain('set initial input of cfg to "npm run dev\\n"');
+    const cmdIndex = script.indexOf("npm run dev");
     const splitIndex = script.indexOf("paneRightCol to split");
     expect(cmdIndex).toBeLessThan(splitIndex);
   });
@@ -1342,14 +1333,13 @@ describe("generateTreeAppleScript", () => {
     expect(script).toContain("set environment variables of cfg to");
   });
 
-  it("shell-quotes multi-argument root pane command", () => {
+  it("preserves multi-argument root pane command text", () => {
     const plan = makePlan(
       { type: "pane", name: "server", command: "npm run dev" },
     );
     const script = generateTreeAppleScript(plan, "/tmp/project");
 
-    // The command name stays unquoted, but arguments are shell-quoted
-    expect(script).toContain("input text \"npm 'run' 'dev'\" to pane_server");
+    expect(script).toContain('input text "npm run dev" to pane_server');
   });
 
   it("auto-resize only applies to first root-level right-split", () => {
@@ -1467,6 +1457,7 @@ describe("status trap", () => {
     const script = generateAppleScript(plan, "/tmp/test", null, undefined, "myapp");
     expect(script).toContain("trap");
     expect(script).toContain("myapp.active");
+    expect(script).toContain("myapp.pid");
     expect(script).toContain("EXIT HUP");
   });
 
@@ -1499,6 +1490,13 @@ describe("status trap", () => {
     const script = generateTreeAppleScript(plan, "/tmp/test", null, undefined, "myproject");
     expect(script).toContain("trap");
     expect(script).toContain("myproject.active");
+    expect(script).toContain("myproject.pid");
+  });
+
+  it("writes the root shell pid sidecar when projectName provided", () => {
+    const plan = planLayout();
+    const script = generateAppleScript(plan, "/tmp/test", null, undefined, "myapp");
+    expect(script).toContain(`printf '%s\\\\n' \\"$$\\" > \\"$HOME/.config/summon/status/myapp.pid\\"`);
   });
 });
